Анимация таймера. Доработка
