+++ source/scss/blocks/game.scss
@@ -78,6 +78,7 @@
   font-size: 4.8rem;
   line-height: 0.906;
   letter-spacing: 0;
+  min-width: 2em;
 
   @media (min-width: $stop-scaling) {
     font-size: 48px;
